refactor(apple): always use Route for routing destinations, improve router
