+++ apple/TasteNotes/UI/Components/BarcodeScannerView.swift
