+++ apple/TasteNotes/UI/Components/RatingView.swift
