+++ apple/TasteNotes/UI/Screens/ProductScreenView.swift
@@ -64,7 +64,7 @@ struct ProductScreenView: View {
           .foregroundColor(.primary)
 
         HStack {
-          NavigationLink(value: product.subBrand.brand.brandOwner) {
+          NavigationLink(value: Route.company(product.subBrand.brand.brandOwner)) {
             Text(product.getDisplayName(.brandOwner))
               .font(.system(size: 16, weight: .bold, design: .default))
               .foregroundColor(.secondary)
